Use 3.10 with cython
3.11 seems to use to require more stringent cython.

While this is a good thing. Right now we should focus on one issue at a
time.

The issue here is TOM, not 3.11 Cython stringencies

+++ mathics/core/number.py
@@ -70,7 +70,7 @@ def _get_float_inf(value, evaluation) -> Optional[float]:
 
 def get_precision(
     value: BaseElement, evaluation, show_messages: bool = True
-) -> Optional[float]:
+) -> Optional[int]:
     """
     Returns the ``float`` in the interval [``$MinPrecision``, ``$MaxPrecision``] closest
     to ``value``.
